# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BuddyBook API",
    "version": "1.0.0",
    "description": "สร้างอัตโนมัติจาก router ของ Express (apps/api/src/lib/openapi.ts) — รายละเอียด body ดู zod schema ใน controller และ API_Endpoints.md"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "internalToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-token"
      }
    }
  },
  "paths": {
...
    "/api/v1/docs": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "GET /api/v1/docs",
        "description": "สาธารณะ",
        "security": [],
        "x-access": "public",
        "responses": {
          "200": {
            "description": "สำเร็จ"
          },
          "400": {
            "description": "ข้อมูลไม่ผ่านการตรวจสอบ (zod)"
          }
        }
      }
    },
...
    "/api/v1/openapi.json": {
      "get": {
        "tags": [
          "openapi.json"
        ],
        "summary": "GET /api/v1/openapi.json",
        "description": "สาธารณะ",
        "security": [],
        "x-access": "public",
        "responses": {
          "200": {
            "description": "สำเร็จ"
          },
          "400": {
            "description": "ข้อมูลไม่ผ่านการตรวจสอบ (zod)"
          }
        }
      }
    },
...
    "/api/v1/webhooks/stripe": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "POST /api/v1/webhooks/stripe",
        "description": "สาธารณะ",
        "security": [],
        "x-access": "public",
        "responses": {
          "200": {
            "description": "สำเร็จ"
          },
          "400": {
            "description": "ข้อมูลไม่ผ่านการตรวจสอบ (zod)"
          }
        }
      }
    },
...
  }
}